requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
configparser==5.3.0
decorator==5.1.1
Django==4.1.5
django-bootstrap-v5==1.0.11
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
importlib-metadata==6.0.0
ipython==8.7.0
jedi==0.18.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyMySQL==1.0.2
pytz==2022.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
